requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.6
pylint==2.9.6
pylint-quotes==0.2.3
autopep8==1.5.7
python-dotenv==0.19.0